+++ packages/router/src/adapters/subgraph/index.ts
@@ -1,32 +1,11 @@
-import { BigNumber } from "ethers/lib/ethers";
 import { GraphQLClient } from "graphql-request";
 
-import { ActiveTransaction, SingleChainTransaction, SubgraphSyncRecord } from "../../lib/entities";
-import { ContractReaderNotAvailableForChain } from "../../lib/errors/contractReader";
+import { ContractReader } from "../../lib/entities";
+import { ContractReaderNotAvailableForChain } from "../../lib/errors";
 import { getContext } from "../../router";
 
 import { getSdk, Sdk } from "./graphqlsdk";
 import { getActiveTransactions, getAssetBalance, getTransactionForChain, getSyncRecord } from "./subgraph";
-
-export type ContractReader = {
-  getActiveTransactions: () => Promise<ActiveTransaction<any>[]>;
-  getTransactionForChain: (
-    transactionId: string,
-    user: string,
-    chainId: number,
-  ) => Promise<SingleChainTransaction | undefined>;
-
-  /**
-   *
-   * Returns available liquidity for the given asset on the TransactionManager on the provided chain.
-   *
-   * @param assetId - The asset you want to determine router liquidity of
-   * @param chainId - The chain you want to determine liquidity on
-   * @returns The available balance
-   */
-  getAssetBalance: (assetId: string, chainId: number) => Promise<BigNumber>;
-  getSyncRecord: (chainId: number) => SubgraphSyncRecord;
-};
 
 const sdks: Record<number, Sdk> = {};
