+++ packages/router/src/lib/entities/contractReader.ts
@@ -1,6 +1,27 @@
 import { CrosschainTransaction, TransactionData } from "@connext/nxtp-utils";
+import { BigNumber } from "ethers";
 
 import { TransactionStatus as SdkTransactionStatus } from "../../adapters/subgraph/graphqlsdk";
+
+export type ContractReader = {
+  getActiveTransactions: () => Promise<ActiveTransaction<any>[]>;
+  getTransactionForChain: (
+    transactionId: string,
+    user: string,
+    chainId: number,
+  ) => Promise<SingleChainTransaction | undefined>;
+
+  /**
+   *
+   * Returns available liquidity for the given asset on the TransactionManager on the provided chain.
+   *
+   * @param assetId - The asset you want to determine router liquidity of
+   * @param chainId - The chain you want to determine liquidity on
+   * @returns The available balance
+   */
+  getAssetBalance: (assetId: string, chainId: number) => Promise<BigNumber>;
+  getSyncRecord: (chainId: number) => SubgraphSyncRecord;
+};
 
 export const CrosschainTransactionStatus = {
   SenderPrepared: "SenderPrepared",
