+++ packages/router/src/lib/entities/contractWriter.ts
@@ -1,0 +1,27 @@
+import { PrepareParams, RequestContext, FulfillParams, CancelParams } from "@connext/nxtp-utils";
+import { providers } from "ethers";
+
+export type ContractWriter = {
+  prepare: (
+    chainId: number,
+    prepareParams: PrepareParams,
+    requestContext: RequestContext,
+  ) => Promise<providers.TransactionReceipt>;
+  fulfill: (
+    chainId: number,
+    fulfillParams: FulfillParams,
+    requestContext: RequestContext,
+  ) => Promise<providers.TransactionReceipt>;
+  cancel: (
+    chainId: number,
+    cancelParams: CancelParams,
+    requestContext: RequestContext,
+  ) => Promise<providers.TransactionReceipt>;
+  removeLiquidity: (
+    chainId: number,
+    amount: string,
+    assetId: string,
+    recipientAddress: string | undefined,
+    requestContext: RequestContext,
+  ) => Promise<providers.TransactionReceipt>;
+};
