+++ packages/router/src/lib/entities/index.ts
@@ -7,6 +7,7 @@ export {
   FulfillPayload,
   CancelPayload,
   SubgraphSyncRecord,
+  ContractReader,
 } from "./contractReader";
 
 export { PrepareInput, PrepareInputSchema } from "./prepare";
@@ -14,3 +15,7 @@ export { PrepareInput, PrepareInputSchema } from "./prepare";
 export { FulfillInput, FulfillInputSchema } from "./fulfill";
 
 export { CancelInput, CancelInputSchema } from "./cancel";
+
+export { Cache, GetOutstandingLiquidityParams, StoreOutstandingLiquidityParams } from "./cache";
+
+export { ContractWriter } from "./contractWriter";
